Store metadata within the EEPROM map
To expand the number of ways that we can select an EEPROM map (e.g. by
manually by name), we need to store some metadata (e.g. name) with the
EEPROM map. So extend the eeprom ops structure to keep these data.

Also rename map ops structure to emphasize it new nature and finally
constify it.