Add support for accessing device via I/O memory
Add another optional (and enabled by default) connector, which uses an
explicitly specified start location of the device I/O memory to map it
into the process memory in order to provide access to the device
registers.

This of accessing the device can be useful in cases where libpciaccess
is not available (e.g. for debugging on embeded devices).