Rename EEPROM map Default to 5416
This map is by default used by the ath9k driver, which supports only
.11n chips, but this map is not default for other chips (e.g. AR5xxx).
Therefore, rename it to avoid confusion. The new name is 5416, since
AR5416 was the first chip that uses this EEPROM layout, but acually this
map is suitable for AR9160 and AR92xx chips too - note this fact in the
description.

No other functional changes besides the files and structures renaming.