Correct AR5416/AR9285/AR9287 endians handling
Native format of the AR5416 EEPROM and its friends (AR9285, AR9287) is
little-endian, but the multi-byte fields in these EEPROMs can be stored
in big-endian format. There is a dedicated flag in the EERPROM data,
which indicates the endianness of multi-byte fields.

On the other hand, the code, which fills the EEPROM data structure,
detects the EEPROM endianness by checking the MAGIC field. If the
formats do not match then it swaps bytes in each EEPROM word and assumes
that multi-byte fields should be swapped as well. This breaks the EEPROM
processing, if its words are stored in a one endianness, but the process
is run on a host with another endianness (e.g. parsing on a x86 host a
binary dump, which is taken on a big-endian router).

Fix this by tracking a host endianness, explicitly checking the
endianness flag of EEPROM, and performing the multi-bytes fields swap
only if formats do not match.