eep5416/9285/9287: unify data size handling
For each EEPROM map add a couple of macros that defines the EEPROM data
start location and a size of this data. And unify they usage across all
maps to make code clear and similar.

Also fix the customer data printing. Each EEPROM map has a customer data
region of different sizes, but they all print exactly 64 bytes, what
looks like a copy-paste from the AR5416 code. Fix this using the size of
the customer data array to determine region to print. And explicitly
define the customer data size for each map using a macro to avoid magic
numbers usage and to make it clear that they are all different.