Use an intermediate buffer for EEPROM dump
Load the EEPROM contents into the intermediate buffer first and only
then start the content parsing. All modifications, like the byte
swapping of field values, are performed using the parsed EEPROM data,
while the buffer contents stay unchanged. So the buffer is used as a
cache. This approach will be useful for several future utility options,
such as saving EEPROM data in a file and supporting of a EEPROM content
modification.

Only the AR5416/AR9285/AR9287 EEPROM maps were converted. AR9300 needs
more work due to the complexity of the data loading code.