eep5416/9285/9287: separate ver and csum checks
Separate the checks of the EEPROM version and the checksum and first
verify the EEPROM version. Since, if the EEPROM data is completely
corrupted, the version field will contain sensless information, and its
printing is more informative than the message about a bad checksum. Also
make the EEPROM version error message more informative by adding to it
the revision information and the raw field data to it.