Add generic EEPROM I/O word byteswapping
Add an intermediate layer for EEPROM I/O with the word byte swapping
support and use it to load the EEPROM contents into the buffer with the
correct (little) endians at the EEPROM filling stage instead of
performing the swapping later during the EEPROM check operation. Byte
swapping requirements are detected by checking the Magic value as
before.