Add EEPROM content saving support
It is useful to preserve the original EEPROM contents before any
modifications or save it for future "offline" analysis.

Add an action that saves the data, which was fetched from the EEPROM.
Due to the fact that this action uses the intermediate buffer as a data
sourse for writing, saving EERPOM whose map does not use buffer (e.g.
AR9300) is impossible.