Fix chip version checking macros
The macros for testing for AR9285/AR9287 chips have an mispelled
argument ('_ah' instead of 'edump'), which does not match the
instructions inside macros. But everything works well, since each
function that uses these macroses has a variable called 'edump'.

Fix this issue and prevent such cases in the future, using the argument
name that has the underscore prefix for the each macros of version
checking. And drop odd brackets as well.