Add EEPROM unlocking support
A lot of EEPROM ICs have a WriteProtection input which is controlled by
a GPIO of the wireless chip to prevent accidental writing. This
WriteProtection input can be connected to any GPIO, but most WLAN card
vendors follow the reference design and connect it in the same way as
recomended by Atheros.

Add support for the writing unlocking via GPIO so that the utility can
automatically unlock writing when it needs to update EEPROM contents.