Rework section selection for dumping
Not every EEPROM map supports dumping of each section (e.g. AR9300 does
not support power information printing). But the current code requires
existance of a function for each section. Which leads to creating of a
stub functions.

Group pointers to the dumping functions of the EEPROM section into an
array and allow the pointer to be NULL to indicate that the dumping of
this section is not supported by this map.