Rework chip name detection
The name of the chip of the AR5xxx series depends not only on the
silicon version, but also on the silicon revision number. This is
especially important for determining the names of legacy chips. So,
split the routine of the chip name detection into two routines: one for
AR5xxx names detection, and another for  AR9xxx names detection. The
first uses both version and revision subfields, and the second uses only
the version subfield.

Also correct few SREV constant names:
* AR5416_PCI -> AR5416,
* AR5416_PCIE -> AR5418.