Add an option to erase a CTL data
Add an option (EEPROM update param) that triggers the CTL (Conformance
Test Limit) data erase.

CTL can cause unexpected limitations in situations when you are allowed
to operate at full power in full range, and a card itself is calibrated
for full power, but CTL forces you to follow the rules for unlicensed
users. So having an option to bypass this is highly welcome :)

Only AR5211 EEPROM is suported for now.