eep9300: use intermediate buffer for OTP too
During OTP processing use the same data handling approach as was used
for EEPROM processing: read the OTP memory contents to the intermediate
buffer and then interact only with them.

This change unifies access to various memory and finally allows to dump
OTP contents to file since the file writing code uses intermediate
buffer as the source of data.