eep9300: rework blocks processing
Factor out blocks processing loop to separate function and call it
instead of using a one way 'goto'. This allows us to really try to parse
several regions instead of just checking is region begin is blank or
not.

This change is a preparation to support processing big-endians dumps
on the little-endians host and vice versa.