eep5416: modal: rework line printing macros
Rework line printing macro in such way that user now should provide a
callback macro that will prepare data in the string buffer instead of
providing of a format string. The new calling method is more flexible
and gives us a reach abilities to format data.

This change is a preparation for upcoming per-chain data compaction.