eep5416: add support for CTL erasing
Add support for CTL (Conformance Test Limit) erasing of
AR5416/AR9160/AR92xx EEPROMs. As stated earlier during the option
introduction this is helpful in cases when you have a *LICENSE* to
operate station with a greater power levels, but prefer to use a stock
driver.

Since this feature is very dangerous it was hidden under the special
build time option.