eep5416: fix power detector curve data selection
Power detector calibration data are store without any gaps. E.g. if some
gain is disabled (by the gain mask) then data for next enabled gain will
be placed immediatly in the per-gain calibration data array. I.e. if we
have a gain mask = 0x06 then data will be placed at the 0 and 1 elements
of array. But earlier the dumping code assumes that position in the
array is equal to gain mask bit position. What leads to missing a
correct data and to printing a zeros instead.

Fix this by reworking the dumping PD calibration code to correctly
select a per-gain data.