eep9300: rework bytestream extraction approach
EEPROM is stream of 16-bits words in the Little-endian format, while OTP
is byte stream, so it utilizes Native-endians. Earlier we explicitly
convert OTP data to 16-bits Little-endian words and use common
bytestream extraction routine that parse internal buffer assuming it
always contains Little-endian words. But such approach mangles OTP dump
on a Big-endians machine: bytes in each 16-bits word become exchanged.

Dump stays parsable since we use EEPROM handling code for OTP dumps too
and EEPROM tries both direct and byteswapped readings. But anyway we
loose ability to take a precise dump and dumps of a same card differ if
they are taken on machines with different endians, what do not make life
easier.

In order to fix the OTP dumping byteswaping issue, store OTP data in the
internal buffer as is without endians touching. Also introduce a flag
that indicates to the bytestream extraction routine the endians of data
in the internal buffer.