Tree-wide rename of EEPROM data loading functions
Rename EEPROM data loading operation from fill_eeprom to load_eeprom as
well as functions from all maps that realize this operation. Old name
was selected in a such time when this operation really parses compact
EEPROM data representation and fill a "eepom" structure. While the new
name indicates that data will be loaded from a EEPROM. This renaming is
a preparation for upcoming separation of EEPROM and OTP memory loading
code.