Make EEPROM loader optional
We are goin to introduce a dedicated OTP memory data loader, which
could be called in case of empty EEPROM and available OTP. Moreover,
some modern chips have no EEPROM at all.

So, rework EEPROM loader invocation in a such way that it become
optional and it will be called only in case of existing EEPROM access
interface, which could be provided by a connector or selected according
to chip identifier.