eep5416/9285/9287: fix Big Endian EEPROM data loading
Previously, we assumed that the magic field is always in Little-endian
(LE) format, but in fact it is formatted in the same way as all other
multibyte EEPROM data fields. E.g. if an EEPROM data is in Big-Endian
(BE) form, then the magic field will also be stored as a BE number. We
use the magic number not only to determine the overall validity of an
EEPROM data, but also to decide whether we should activate I/O
byteswapping or not (e.g. to handle an LE formatted EEPROM on a BE CPU).
So if we have a truly BE formatted EEPROM data, we will mangle it by the
I/O byteswapping. Moreover, some fields are single-byte, and
byteswapping not only mangles data within multibyte fields, but also
swaps adjacent single-byte fields.

The EEPROM contains a special flag (in the eepMisc field) that indicates
that all multibyte fields are BE formatted. We use this flag to swap
bytes of multibyte fields _after_ data loading (with possible I/O
byteswapping) to turn those multibyte fields to native-endian. But due
to the swapping of adjacent fields during data loading, we actually
theat the 5GHz support flag of the adjacent opCapFlags field as the BE
format indicator, what only amplifies the data mangle.

The chip driver knows a priori whether an EEPROM data is LE or BE. This
is because BE enconding is only used when a data is actually stored in
flash of an access point that has a BE CPU. But a dumping utility should
be able to parse a data without a priori knowledge (i.e. it should
detect the byte order and automatically toggle byteswapping).

Since we are unable to reliably detect BE data using only the magic
field and BigEndians flags, we should use some heuristic approach to
make a more accurate decision about the I/O byteswapping requirement.

Use the calibration software version field for this purpose. A normal
content of this field has a structure that allows us to distinguish
between byteswapped LE data and truly BE data. In fact this approach
allows to detect even byteswapped BE data, however I never saw such
case.

Along with the BigEndians flag, this increases the probability of the
Big-endian data detection. This solution is not 100% reliable, but it is
better than incorrect decoding of 100% of Big-endians flash dumps :)

If version field test does not allow to determine byteswapping presense,
then code fall back to default mode and assumes that a data is LE
encoded. This should reduce false positive byteswapping detection.