eep5416/9300: avoid checking SREV in subfield mask macro
Latest chips (e.g. AR9462, AR9565) have more GPIO lines and so require
a bigger rfSilent (RF Kill) GPIO line number subfield. Earlier we define
a macro with subfield mask, what checks SREV and use bigger mask for
latest chip. This is the same approach that is utilized in the device
driver. But the device driver always have an access to a chip, while
the utility usually have only access to dump of EEPROM contents. So when
parsing dump of EEPROM of a modern .11n card the utility could truncate
MSB(s) of rfSilent GPIO line number.

So define a dedicated set of macros for AR9300 eeprom map with extended
GPIO line number subfield mask. And remove SREV checking in the subfield
mask of previous EEPROM formats and constantly use reduced subfield
mask.

MS bits truncation does not happend at the moment since we print a whole
rfSilent field content as is without any parsing. So this change mostly
fix a code readability.