pci: fix AR5418/AR9220/AR9227 chip names
Theare no such chips as PCIe AR5416 or PCI AR9280 or PCI AR9287. In fact
AR5416 with PCIe interface is known as AR5418, AR9280 with PCI interface
is AR9220 and AR9287 with PCI interface is known as AR9227. Account this
information in the PCI device name printing to be more precise.