Add common buffer for unpacked data
It is become quite common for modern chips to keep calibration data in a
packed form. Particular data format handlers unpack them before usage,
but use a private buffer for data keeping and modify data in some way
(e.g. harmonize endian with host CPU). Data modification make data
mangled in some way, resulted data a differ from unpacked data.

We want to be able to save unpacked data as is for utility debugging and
to be able to analyze unpacked data with 3rd party software (e.g. with
hexdump :)). So introduce a transient buffer that could be used to store
unpacked data and a user action (comand line argument) to save buffer
contents to a user specified file. This buffer behave like buffer for a
raw readed data: format (map type) descriptor specify maximal buffer
size, utility allocates memory for it, during the loading stage unpacked
data are saved to the buffer and real data length is specified in the
additional common field, after that utility could save unpacked data if
user requests so.