eep9300: rework template fetching
Pass the pointer to the template EEPROM lookup function to the data
decompression routine to make it particular template set agnostic. Also
turn template pointer from eeprom structure to uint8_t with the same
purpose to make function EEPROM type agnostic.

This is a preparation for making decompression code common.