eep9300: rework compressed data passing
We already pass unpacked header with compression parameters to
the decompression routine, so there are no reasons to pass compressed
data with the header part.

Skip header part and pass only comressed data to the decompression
routine to avoid pointer shift each time we need to obtain the
compressed data.

Also constify input function arguments to make it clear that this data
will not be changed.