Add Tx Power defining macro for EEPROM templates
Tx Power is usually stored in a 1/2 dBm units. So a filled EEPROM
template could looks quite stranger: we saw 24 in a Tx Power field
(e.g. in target tx poewr), but in fact value of this field mean only
12 dBm.

In order to make template reading comfort add a simple macro that
convert power value to the storage value (i.e. multiply value by x2).
And use this macro to fill AR93xx templates to make them a bit more
simple for reading.