Support for autonomous actions
Essential part of the utility is to handle input data and optionally
interact with hardware to fetch this data. So the utility have an
advaced common interface for data fetching (connectors) and connector
configuration is always mandatory. This is due to the fact that
connector descriptor accessed via pointer in many places without
checking it for validity.

But sometime we need an autonomous (self-contained) action that do not
interacts with outer word, e.d. do not access hardware, do not read
input dumps, etc. Require the user to specify a connector (see above),
that will be in fact ignored is senseless. Reworking whole code and
check connector pointer each time before the access is not an option.

So, add a special action flag that indicates that a particular
action is autonomous. If the user has not specified a connector
invoking such an autonomous action, then silently setup a special
'stub' connector. This 'stub' connector actually do nothing and its
purpose to make init and exit handlers invocation safe without rewriting
a lot of code.