Rework connectors usage string generation
We are going to add another one connector. As a preparation for them,
rework connectors usage string creation in the build time. Switch from
large cascade of #if/#elif/#end to a set of independently defined
per-connector macros, each of which contains usage string for item
connector. And then join these macros together to construct the usage
string. If some connector is not configured then define coresponding
macro as an empty string ("").

This reduce duplication and should make code readable.