eep5416: make power detector dumper code more clear
Rename 'pwr' and 'gain' local variables to 'pwridx' and 'gainidx'
respectivly. This should clearly indicate that these variables are for
array indexing purpose. Also new names should help to avoid confusion
of power and gain *indexes* with *values*, what is the most important
goal of this change.

It is possible to rename these variables to typical 'i' and 'j', but
code is quite long complex, so reader could easly lost purpose of each
one.

This is only renaming of a couple of local variables without any
functional changes.