eep5416: prepare power detector dumper for generalization
Power detector dumper merges multiple (per PD feedback gain) curves to a
structure with a common power axis. This is done to make output nicely
looking. The size of this common structure depends on number PD gains
and on number points of calibration (icepts). At the moment both
parameters are constant. But generilization of dumping code require to
make these parameters variable. This is lead to the variable size of
structure field, which is acceptable by gcc, but clang/llvm do not like
such code style and promise to never support this.

So to prepare for code generalization and to stay clang/llvm compatible,
split structure with multiple curves and embedded power axis on to two
arrays. One of each keeps power axis and the second one keeps all
per-gain detector curves.