eep5416: power cal: split power and detector arrays
The sizes of output power and detector data sets are depends on number
of power detector gains and number of calibration points. Each of these
number are different for different EEPROM formats. So pass these sets
independently to per-chain & per-frequency data dumping routine. On one
hand we could still pass them together and split on to separate sets in
the routine itself. But top level code anyway should know sizes of both
sets, so split sets earlier.

This is a next step in the generalization of power detector data dumper,
what we should perform to use it for other (AR9285/AR9287) EEPROM
formats.