eep9880: support 'generic' template
The OTP memory contains only a diff against some predefined calibration
data template. There are several templates, so OTP also contains a
template id to instruct the decompressor which one should be used.
Absense of a required template causes a decompression failure.

The 'generic' template (ref = 1) was not added previously as I was not
aware of any card that uses it in the wild. It looks like the Compex
WLE900VX miniPCIe card [1] is one of them and it triggers the
decompression failure:

  Found block at 0: comp=3 ref=1 length=XXX major=X minor=X
  can't find reference eeprom struct 1

In order to fix this, add the 'generic' template to the utility so that
it can decompress the OTP contents of this and possibly other earlier
produced cards.

Refs: gh-6

1. https://compex.com.sg/shop/wifi-module/802-11ac-wave-1/wle900vx-2/