RAW data reading mode framework
Sometime the utility unable to check EEPROM/OTP contents. In particular,
this could be caused, for example, by a conntents corruption or a missed
template in the decompression algorithm. The utility declined any
handling of data if they looked inconsistent. The readed data saving
will be rejected as well in case of a failed check.

But sometimes it is desireable to do just our best and at least save the
extracted data as it is. E.g. for a partial audit, data consistency
recovery, or simply for further utility development in a relaxed and
more comfortable environment. So add a capability to save the read RAW
data as-is. Without any checks and even without I/O byteswapping.

This change introduces only the RAW data saving framework: new commands,
new code execution paths that bypass checks, and a set of flags
indicating whether a particular format handler is capable to fetch data
without any preprocessing. Per-format support for the RAW data loading
will be introduces by the following changes.

Refs: gh-4, gh-6