eep9300: RAW EEPROM/OTP reading support
Add support for RAW EEPROM and OTP contents loading if it is requested
by the caller. It assumes complete contents ignoring, including I/O
byteswapping check bypassing and decompression avoidance.

A WNIC can contain two data sources: EEPROM and OTP, with EEPROM being
tried first in the common case. So the EEPROM RAW reading code has been
augmented with empty/missed EEPROM checks to skip the EEPROM and
automatically continue with OTP memory as the data souce. The user can
still explicitly specify which memory should be read using the
appropriate utiltiy command: saveraweep or saverawotp.

The change was runtime tested with QCA9565 based card. The card contains
data only in the OTP memory, but the EEPROM reading should be reliable
as well.

Refs: gh-4