eep5211: RAW EEPROM reading support
Add support for RAW EEPROM contents loading if it is requested by the
caller.

The RAW data loading assumes complete contents ignoring. So, in this
mode, the loader will read maximum amount of data (16 KiB at the
moment). As a side effect, read data may contain multiple repeated
copies of the EEPROM due to the cyclic nature of the read operation.

The change was runtime tested with AR5414 chip.